Make the plugin compatible with the asset-pipeline plugin (required for Grails 2.4).
Additionally the jQuery and Bootstrap files were moved to a subfolder
called "platformTools" to avoid clashing with application files
that have the same name.

When this is the case, the application file wins.
Because the application might use a newer version of Bootstrap,
this would possibly lead to the design being broken for the Platform Core views.